ref: cleanup algebra plugin templates - part 1 (#711)
This adds a struct that contains all algebra definitions to avoid going via namespaces, and also so that not everything is accessed through the transform3, making the algebra type semantics less confusing. Especially in the tests there is, however, still quite some amount of code using the globally defined scalar. Also, in places where default parameters were needed, I put a macro. So this will need a follow up