Add metrial map display to svgtools (#720)
Adds material a maps display to detray svgtools. The material maps will be drawn with the detray_detector_display tool alongside their surfaces, if a material maps file was provided. Three different color maps are currently available, but need to be set in c++ for now.

Also, fixes the proto surface type, so that passive surfaces can be drawn in a different color than sensitive surfaces, e.g. the toy detector beampipe is now grey.