feat: cuda navigation validation (#732)
Run the navigation validation on device with data that was gathered during navigation run on the host. This adds a navigation state view so that memory allocations needed for the navigation inspectors can be handed through. The navigation object tracer, which records data whenever the navigator finds an object of given type, now also records the global track position and direction. This can now be written to file as well, and compared between host and device.

In order to display the new truth and recorded intersections traces, the svg code had to be adapted a little. The CUDA navigation validation is run in the CI now, however, I did not know how to get the truth data there, so I recreate it, but with less statistics.

I also discovered that the current estimation of the maximum number of candidates has a bug, which I fixed by using a functor to query the every acceleration structure according to the link the volume provides.