Add material maps generator (#709)
Moves the generation of the material maps from the detector helper to a dedicated generator. The material maps generator is used in the new toy detector to generate material maps on the portals of those volumes that contain sensitive modules and the beampipe. Since the old toy detector also depended on this code, it was removed. The detector helper was cleaned up and integrated into the wire chamber creation, since it should not be used anymore for any new detector creation.

Edit: Also fixes an issue with the surface range that is returned from the surface factories