Refactor test library and add CUDA test lib (#735)
This PR adds a tool for the CUDA navigation validation, so that any detector read in from json can be checked. For this, I made new test libraries for CUDA and SYCL that can be shared between the detray tools and the integration test that are run in the CI.

This also refactors the python implementation, so that it calls the validation tools to generate the plot data first and then plots it. For now, only the ray scan plots are re-implemented, more to come.