feat: Add additional track posistion and residual plots (#740)
Added a method to write the track positions and directions that were recorded whenever the navigation has reached a surface to csv. The csv file is read back into the python plotting tools where 2D track position and residual plots are prepared from them (e.g. host vs. device). I moved the csv IO functionality in python into the same files where the corresponding plotting functions are defined.

Also fixed the selection of the number of tracks to check in the validation tools and made sure that the python code is handling the data in double precision, so that the minute residuals between host and device code become visible at all (but only when removing the fast math option for CUDA and even then, it is quantized to multiples of the machine epsilon)