feat: material validation in navigation (#741)
Traces the material per track as discovered by the navigator. Implemented both for host and device. The material tracing has been added to the navigation validation, where it will log the amount of material every ray or helix has encountered. Apart from that, a new test has been added (material_validation) which will run its own version of material tracing with the full covariance transport including material interactions enabled. After accumulating the material, it will compare against the material scan, which can now be pinned to the test whiteboard. Both, material scan and navigation material trace are dumped into csv files.

The material validation test has been added to all testbed detector CI test for both host and device, as well as to the detray_material_validation tool, which can be run for any detector read from json. A dedicated detray_material_validation_cuda exists, too, which allows to run the navigation material tracing on device.

Also fixes a bug, where the material scan was logging material of a dummy intersection and tries to prevent the double counting of material on matching portals of neighboring volumes. The latter could not fully be tested, however, since the material maps in the toy detector are not linked to the portals of gap volumes (this is why the toy detector material CI test sees too little material at the moment)