Fix non-conformant use of iterator type aliases
The `check_connectivity` function in the detector scan utilities header
makes use of the `difference_type` typedef on the iterator of a vector,
but this is not defined by the standard, and I was running into
compilation errors stemming from this. The correct solution to this is
to use the `std::iterator_traits` class (because an iterator can also be
a raw pointer which doesn't carry any type aliases) but since we are
using a vector I simply went ahead and used the
`std::vector::difference_type` definition.