Migrate static strings to `std::string_view`
You wouldn't guess it, but it's yet another case of NVCC not liking
detray code. As usual, detray hasn't actually done anything wrong, but
some code in detray is leading to segmentation faults in targets built
from translation units compiled with NVCC.

The reason this happens is that detray make extensive use of names and
tags in the form `inline static const std::string`. This is perfectly
legal, but the mechanics of this mean that the compiler has to inject
some code at the start of the program which takes the string literals
and stores them in memory. This is because the compiler is not allowed
to simply put the strings in the `.data` section of the object, they
have to be properly constructed and they have to live in `.bss`. In the
meanwhile, the `inline` keyword requires the linker to deduplicate the
string across translation units _and_ to make the requisite updates to
the initialization code. I suspect that this is where things are going
wrong, because the initialization code is throwing a segmentation fault.

Note that for whatever reason, the old code works fine in C++17 mode,
but breaks in C++20 mode. I have tried to create a minimal reproducer
for the CUDA team, but I haven't managed so far.

This commit replaces the `inline static const std::string` types with
`static constexpr std::string_view` types, which fix this issue and
should also work a bit better (e.g. they are constexpr, allowing more
optimization to happen, and they don't need construction).