Write out data with momentum information and clean up plots (#760)
Adds the momentum range to the data file names that are generated during the navigation validation, so that the validation can be run for different momenta automatically. Also cleans up the corresponding plots (font sizes etc.)

Fixes a bug, where in the svg the intersection positions on concentric cylinders were not displayed correctly after reading them back in from file, since the files only safe the bound coordinates. Therefore, the bound_to_global method is now used in the illustrator to get the correct positions, setting the first intersection always into the origin by convention (this is a dummy record that captures the initial track parameters).