Rename volume and surface classes (#761)
Pure renaming of surface and volumes classes to align better with ACTS and thesis writeup. In principle, it should become visible from the class name that these are not purely geometric classes. The volumes are surface containers with acceleration data structures used in the navigation and the surfaces offer tracking specific functionality to the user (which also compile-time couples them to our choice of track parametrizations).