Avoid using `std::sort` in device code
Although C++20 makes `std::sort` a constexpr function, that doesn't
automatically make it safe to use in device code. In particular, I've
been running into a problem where the CUDA compiler (or more accurately,
the PTX assembler) cannot statically determine the stack size of
functions using `std::sort`, causing it to fail. This commit removes a
check which allows the use of this sorting function in device code and
always uses the custom, safe implementation.