Add missing eta range in SYCL propagator test
The CI for #771 has been failing mysteriously, and I have found that
this is actually due to the fact that the SYCL propagator tests are
missing a configuration parameter limiting the eta range to [-3, 3] like
the other propagator tests have. This commit adds this missing range and
seems to fix the CI bug.