Remove distiction between free track vector and free track parameters (#792)
The free vector and the free track parameters are pretty much the same now, so let's use the classes public interface to access the data. This leads to some code simplifications in the track helper and the single parameter access

Also fixed a bug, where the CUDA code would not compile due to missing initializers on the mask values.