ref: bound track parameter interface (#794)
This PR wraps the 6-dim bound raw vector behind the bound track parameters behind a class interface, so that the track helper is no longer needed to interpret the data passed to clients. This improves "Information Hiding" and helps to abstract the internal implementation of the track parameters better from its clients.