ref: fixed window navigaiton (#798)
Remove the external candidate buffer and replace with a static candidate array. The default size is set to 10 candidates for now (so far, it was 20). In order to guarantee correct navigation if the cache size is too small to hold all candidates produced by successful intersections, the candidates are sorted while being inserted into the cache, evicting the last element if it is further away from the track position.

The correctness of the navigation is therefore no longer linked to the cache size of the navigation state. It will keep the 10 best candidates in the cache, while any additional candidates will not be cached initially. The other candidated will be handled by the navigator after updating the cache with a new init() call, once it runs out of reachable candidates.

The navigation state now is index based instead of holding iterators to make it easier to copy/copy assign