Add CMake prefix for benchmark builds
Currently, the build for the benchmark CI job in Gitlab is configured in
Gitlab, so any build changes in the detray repository must be
immediately reflected in Gitlab. This is annoying, and can easily be
fixed by adding a CMake prefix for these builds; that way, the build
configuration can be changed from the GitHub side!