Allow different track charges in navigation validation (#820)
Even though the track generator can be configured with any charge, the navigation validation can only be run with q=1 to avoid possible inconsistencies when the trace data is dumped into csv files. This has been fixed by passing the charge as an additional parameter.

Additionally, the charge getters have been removed from the test trajectories (helix and ray), because it is actually unknown there and should not be guessed to be one.