Use ranges algorithms (#848)
Use range algorithms instead of the equivalent from the standard library. This helps readability and will make the code more robust, since ranges applies more compile-time checks. In a few places, the detray code had to be modified to fulfill the concepts that the ranges algorithms require.

For some reason std::ranges::equal fails in comparing the result of the grid bin views in two unittests, while std::equal succeeds. Because of this, it was not replaced with the ranges version.