Use CUDA native functions in math namespace
This commit extends the math namespace to replace the C++ standard
library of floating point math functions (e.g. `atan`) with the CUDA
variants. The benefit of this is that the CUDA versions are better
optimized for GPU execution, and it enables us to speed them up
massively using fast-math mode.