ref: Use the detector builder for the wire chamber geometry (#853)
Use the detector builder to build the wire chamber test geometry instead of deprecated building code that lives directly in the detector. For this, a new wire_layer_generator was added, similar to the geometry generators of the toy detector. The deprecated geometry building code was removed from the detector, except the function set_transform, which is needed in the jacobian validation. This should be removed again, as soon as different geometry contexts can be put into the detector. The different volume builders are now friends of the detector and (at some point) the builder classes should be the only ones that can modify a detector object.

The wire chamber detector can now additionally be built with straw tubes, e.g:

./bin/detray_generate_wire_chamber --straw_tubes