Update the navigation config to the ODD/ITk defaults (#856)
This PR updates the navigation config to the ODD/ITk defaults. This should help to use a more optimal configuration for the ODD and ITk without passing a specific configuration to detray. Not sure, if there is one configuration that will be optimal for all detectors, though, so the telescope detector needs to be configured explicitly now.

The configurations to the propagation also now have to be passed explicitly, which helps not passing the default configuration by mistake.

Edit: while debugging a navigation failure in the telescope with the new default configuration, I added "a last chance" navigation init with loose surface finding constraints that seems to also recover the odd B-field cases.