Reduce size of navigation state (#868)
Reduce the size of the navigation state to 256 bytes (single precision) by reducing the sizes of the involved data types and conditionally removing the local position from the intersection struct. The local position is currently only used in validation code and removing it brings a measurable performance benefit. This PR also removes the is_init flag from the navigator state and returns it from the update call instead. The navigation/propagation heartbeat can be obtained from the states instead.