Simplify random access iterator advancement
I was surprised while profiling the detray code but a surprisingly large
amount of time is spent in the random access iterator advance function.
This is somewhat inefficiently implemented as it is full of if
statements. I have decided to simplify the function, which leads to a
small but noticeable performance improvement of around 4% in the detray
benchmarks.