ref: Generalize propagation benchmark functionality (#404)
Generalize the benchmark functionality to different detectors and actor setups. Also splits some common benchmark functionality into a detray benchmark library, like e.g. the generation of track samples.

The CPU benchmarks have also been switched to use dynamic scheduling for load balancing.

Also adds the charge conjugation operation to the pdg particle, so that the charge hypothesis can be updated when test tracks are generated with randomized charge (this was triggering an assertion in the benchmarks otherwise)