Make plotting script work without running the banchmarks first (#916)
This way the benchmark can be run e.g. from a bash script of manually first, calling the plotting script in a second step (the plotting has a lot of dependencies and may not work everywhere). This mainly changes how the files to be plotted are registered in the python tool and that some additional data needs to be added to the benchmarks context. This will be used to e.g. distinguish the different algebra plugins and to set the plot labels correctly.
Also removed the 'interleave' option again, as that makes mostly sense for microbenchmarks and would need an additional sorting step to work correctly