Fix state tuple type and remove empty states (#940)
This PR ensures that the local state tuple type in the actor chain also contains the states of all observing actors in a flattened tuple. At the same time, it makes sure that empty default states are not included and therefore also do not need to be passed to the propagator. Consequently, the empty states were removed everywhere now.
In the propagator the actor state tuple type is auto deduced, so that we do not depend on the order of actors in the chain anymore (this will simplify the KF in traccc).