Fixed spelling and opted for explicit typing (#960)
In core/include/detray/navigation/intersection/ray_line_intersector.hpp, there is a line which uses auto for type deduction, which was giving me problems with Fastor, because auto was resolving to an expression template and then causing compilation errors further down the line. I changed it to an explicit type (vector3_type) to force evaluation.