Split the surface implementation (#861)
Make the basic geometric surface interface type as concise as possible and instead only add additional functionality around it when actually needed. Also moves a few pieces of functionality to specialized free functions. The tracking_surface interface and functionality does not change and can continue to be used as is in tracking focused projects like traccc.

This is an attempt to reduce internal coupling, because most of the tracking surface functionality is only needed in very few places. It will also allow for non-tracking focused projects to not depend on tracking functionality in any way