fix: counting error and failed assertion (#959)
Fix an error during the counting of the number of missed candidates at a given step and make sure that the assertion that all missed candidates were counted does not trigger for swapped portals.

Also adds a recounting of the inserted dummy records to check the number of missed surfaces independently and passes a meaningful boolean result from the function.