fix: remove overlaps and make envelopes larger (#962)
Enlarges the portal envelope, similar to what is also used in the generic detector in ACTS. This will become even more important with alignment changes, when the contained surfaces move slightly. Also adds the portal envelope to the maximal detector radius to prevent an overlap of the endcap modules with the outer cylinder portals, since this can lead to skipped surfaces. Adding the same to inner radius does not seem to be necessary