feat: navigation validation function (#958)
Add the functionality to run the navigation against any reference trace and later compare as a separate function that can be used in downstream projects. Allows forward and backward propagation with extra actors and provides detailed statistics on the number of tracks with missing or additional surfaces. For tracks with mismatches to the reference trace a detailed debug log will be dumped to file and an SVG of the track with the truth trace in the detector geometry will be created.

In order to provide all of the different configuration options the navigation_validation_config is used everywhere now and in particular the particle hypothesis has been made available to the navigation validation this way. Any dependency on google test was removed from the validation function.