feat: navigation pause (#975)
Allow to pause the navigation when necessary instead of aborting and resurrecting. This makes it easier to understand, when the propagation hit an error state that is unrecoverable and when the loop was broken by an actor to do other work, with the intention of resuming the propagation later on. Specifically, pause keeps the navigation state intact, while abort may not.