feat: abort messages (#979)
Allow to pass additional messages to the navigator abort function that contain the reason for the failure. This will make it easier to debug, because it can be seen right away from where abort was called and why. The reason for the error is also preserved in the print inspector, which makes it easier to print it to the user in tests outside the debug file.