Infrastructure for more bound parameter types
I would like to investigate the addition of different bound parameter
type storage methods, like e.g. ones with square root decompositions and
UDU decompositions. This will require us to add new bound parameter
types (some of which will not have built-in covariance matrices). This
commit adds the necessary changes to make this happen, lifting the
common parts of the track parameter type into a base type and allowing
that to be extended. Code using the bound track parameters are, where
necessary, updated to be templated.