Fix out of bounds error in detector display tool (#994)
During surface construction, the material link was not set such that is_invalid would be true for a surface with no material. This lead to an incorrect material access in the SVG illustrator when trying to display surface material. Also added some additional checks in the illustrator for too large volume and surface indices that can be requested by the user.