Reduce thread divergence in covariance transport
Currently, the covariance transport uses the Jacobian engine which is
templated on the frame type. While this makes the code easier to read
and write, it requires the compiler to duplicate a lot of code for each
of the frame types, and this duplicated code counts as multiple branches
for the sake of GPU execution. Thus, this templating increases the
amount of thread divergence.

This commit refactors the Jacobian engine into smaller parts, some of
which are templated on the frame type and some of which are not. Client
code and then take a more fine-grained approach to branching and improve
divergence.