feat: use mask ranges (#982)
Use multiple masks per surface for most geometries (except the telescope, where this does not make sense). For this, a new index type was introduced (dindex_range and dsized_index_range), that can be used together with the dtyped_index that we use for the masks and does not need more memory that the single index we have used so far. The indexing is now tested way more thoroughly, and where it was needed, also the other index types were corrected.

This lead to a cascade of changes that needed to be done in the rest of the code:

- masks need to be merged to larger masks, in order to find the true extent of a surface with multiple masks (e.g. in order to fit the material maps correctly to a surface)
- The mask store visitors need to be able to handle a mask range, so they now iterate over a subrange on the mask vector
- The toy detector portals are fitted onto a single surface (per volume) where possible, with changes the surface counts in the detector and all affected tests
- The grids explicitly use the dsized_index_range type now (they did so implicitly before as well), which holds a lower index and the number of indices
- The SVG illustrator generates a surface per portal now, so that missing masks become obvious