fix: Material maps display and scale toy material maps (#998)
Fixes the bin index range for material maps in the SVG display and ensures that the material thickness in material maps that are translated in z scales with regard to the origin instead of their midpoint. This changes the material of the toy detector slightly, including more bins for the now fused disc maps, so that the test need to be adapted.