Fix IO file order (#1002)
Depending on the file name, it is currently possible that the grid or material components are built before the geometry, which fails, because they need detector volumes registered in the detector builder. This PR makes sure that the geometry file is processed before the others by adding it explicitly to the components reader.