Fix helix intersector for cylinders with two solutions (#1003)
For multiple masks, the helix intersector must not stop at the first intersection it finds for a mask, since there are two possible intersections which might possibly be on two different masks. If it finds the intersection in the opposite direction first, this will be discarded by the detector scanner, while the intersection in the "along" direction will never be found anymore, leading to many cases where the detector scan with helices will not find any intersection with the cylinder.