Add resooluton plots for local positions (#1009)
Plot the residuals of the local positions of the intersections found by the detector scan and the navigator to make sure, they are smaller than the intrinsic resolution of a pixel detector. Also fixes a precision issue in dfe libs that prevented the output of enough digits to calculate the residual.