Convert accelerator structure loop to runtime
Currently, the accelerator structure lookup loop is unnecessarily
performed at compile time. This means that both the lookup of the link
as well as anything that is done with the link branches based on a
template parameter, but only the lookup needs to do this. This change
turns the code into a runtime loop, allowing the compiler to deduplicate
the code that uses the accelerator link.