ref: Optimize the portal cylinder intersector (#1053)
Simplified the math for the portal cylinder intersector. Basically, we just need the z and phi coordinates of the point where the navigation ray's perpendicular component equals the radius of the cylinder. This is valid because the cylinder cannot be shifted or rotated. Also fixes a typo in the cylinder unittest