feat: Implement dedicated jacobian for the concentric portal cylinders (#1054)
Implement a dedicated jacobian for the concentric portal cylinders, which use a (phi, z) local coordinate system instead of (r phi, z) in order to perform the grid lookups correctly. Since the cylinders are not rotated or translated and closed in phi, no reference frame is calculated. The path derivative is kept as before