ref: Let the shapes decide the global to local tranform (#1062)
After realizing that the trigonometric functions in the global to local transforms of the different coordinate systems are the most costly part of the intersectors, I removed them as far as possible. Renames the masks to_local_frame to to_local_frame3D for better readability of when we convert into local 3D coordinates instead of local 2D and fixes a bug in the line shape's min_dist_to_boundary method.