Implement Sympy full Jacobian computation
An important bottleneck in detray is the computation of the full
Jacobian from the free-to-bound, bound-to-free Jacobians as well as the
correction terms. This computation involves a lot of matrices with known
substructure, meaning we are performing a lot of unnecessary
computation. This PR implements a Sympy-based generated function for
computing the full Jacobian, which automatically ignores any known zero
or one elements.