ref: separate mask check from intersector (#1063)
This is a prerequisite of dispatching the intersector code by local frame instead of by mask, which will hopefully reduce the branching in the propagation. It will also speed up the portal evaluation, since the underlying surface is only intersected once instead of per mask.

Every intersector is therefore split into the point_of_intersection method, which will do the actual intersection, and the evaluate_mask function, which will perform the "inside" check. In order to pass the resulting data between the two functions, the intersection_point type was introduced, since the intersectors have slightly different requirements in this this regard. Finally, the intersector_base class was introduced to remove some boilerplate code.

The old concentric cylinder intersector has been removed, since it was not used in the navigation anyway.