feat: Distinguish surface intersection edge hits (#1041)
This PR will allow to distinguish whether a surface was hit in the tolerance band of the mask or if the mask was properly hit. For this, the intersectors perform two mask checks on the intersection local position, with and without tolerance, and flag the corresponding enum value for the intersection. For the SIMD backend the enum is replaced with a scalar, so that the boolean mask type that results from the mask check fits the SIMD vector that contains the status information.

This will allow in traccc to only flag holes, if a surface was hit without tolerance, while still feeding the candidates that were found with the loose tolerance setting to the Kalman Filter.