feat: Add multilevel host/device logging to propagation (#1071)
Fixes the logging and adds logging output to the propagation (stepper, navigator and actors). The detector building logging is cleaned up a bit and new logging features are introduced:

- Different log levels: NONE, WARN, INFO, VERBOSE and DEBUG, which can be set in the cmake configuration via the DETRAY_SET_LOGGING flag
- Logging is done via clog and stderr, so that it can easily be piped into a file (except device logging, which has to go to stdout)
- Introduced logging for device: HOST logging uses streaming operators to print detailed object information, HOST_DEVICE logging appears in both compilations and uses printf and DEVICE logging only appears in device code and prints the respective backend (CUDA, HIP, SYCL) that issued the message. Any of the HOST, HOST_DEVICE or DEVICE logging macros can be used anywhere in the code, but will only print messages when appropriate.

The default level is INFO, with VERBOSE adding some messages from inside the propagation loop. DEBUG output is for expert level of information.

The DEBUG logging is turned on in the build-only parts of the CI to make sure it continues to compile.