Remove duplicate surfaces from navigation cache (#1073)
Make sure that surfaces don't appear multiple times in the navigation cache. If the search window is opened on a grid that contains an entire neighbourhood in every bin, a neighbourhood lookup will result in surfaces appearing multiple times in the cache. If it appears to often, the cache can contain only this surface, leading to the navigator getting stuck on that surface.