feat: surface grid svg with neighborhood highlighting (#1075)
Re-establish the issuing of surface grid SVGs where the surfaces that will be seen by the navigation with a given search window are highlighted. Since I had some problems with segmentation faults, I switched the container accesses to range-checked accesses and added some log outputs.
In order to make the SVGs accessible by volume name, I transparently switched out the name_map from the detector to make it bidirectional. In this way, the volume index can now be looked up from the volume name, which is what is needed in order to access any data of the volume in the detector.