Implement generic acceleration structs (#990)
Generalize the acceleration data structure implementation, so that also volume search data structures can be added in the future. In particular, the tracking volume and navigator can now call into surface and volume search data structures. The current surface accelerators are cleaned up and adapted to fulfill the new accelerator concept.

For this, the grid has been split into the 'utility' grid, which is also used by the material maps, and the spatial grid, which is now dedicated to spacial searches of contained surfaces or volumes. Both grid accelerator types (surface and volume) can now be added to the detector accelerator store. The spatial grid is equipped with a dedicated grid collection, which will make it easier in the future to add the mask data of the representing surface (this will need to be added to stay in sync with ACTS).

The actual volume accelerator data structures and their IO will be added in a future PR.