feat: Add the intersector to the spatial grid to mitigate projection effects (#1091)
Implement track projection onto the surface search grids. This is done via straight-line intersection in the spatial grid search method. For some shapes, like the cylindrical shapes, additional information is needed, such as the cylinder radius. For this reason, the corresponding mask was added to the spatial grid, but currently only fully implemented for the cylinder shapes. The mask collection for all grids in the detector is added to the spatial grid collection type, so that is will be copied to the detector automatically. The radius is inferred in the grid builder from the cylinder volume portals